backend/requirements.txt
asgiref==3.10.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.6
Django==5.2.7
django-allauth==65.15.1
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-filter==25.2
django-taggit==6.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
dotenv==0.9.9
et_xmlfile==2.0.0
Faker==37.12.0
idna==3.11
num2words==0.5.14
openpyxl==3.1.5
pillow==12.1.1
pycparser==3.0
PyJWT==2.12.1
python-dotenv==1.2.2
requests==2.33.1
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.6.3
whitenoise==6.12.0
